refactor: move AskYesNo to prompt pkg
