+++ cmd/up/stignore.go
@@ -32,6 +32,7 @@ import (
 	"github.com/okteto/okteto/pkg/linguist"
 	oktetoLog "github.com/okteto/okteto/pkg/log"
 	"github.com/okteto/okteto/pkg/model"
+	"github.com/okteto/okteto/pkg/prompt"
 )
 
 func addStignoreSecrets(dev *model.Dev) error {
@@ -155,7 +156,7 @@ func askIfCreateStignoreDefaults(folder, stignorePath string) error {
 	}
 
 	oktetoLog.Information("Okteto requires a '.stignore' file to ignore file patterns that help optimize the synchronization service.")
-	stignoreDefaults, err := utils.AskYesNo("Do you want to infer defaults for the '.stignore' file? (otherwise, it will be left blank) [y/n] ")
+	stignoreDefaults, err := prompt.AskYesNo("Do you want to infer defaults for the '.stignore' file? (otherwise, it will be left blank) [y/n] ")
 	if err != nil {
 		return fmt.Errorf("failed to add '.stignore' to '%s': %s", folder, err.Error())
 	}
@@ -190,7 +191,7 @@ func askIfUpdatingStignore(folder, stignorePath string) error {
 	}
 
 	oktetoLog.Information("The synchronization service performance is degraded if the '.git' folder is synchronized.")
-	ignoreGit, err := utils.AskYesNo("Do you want to ignore the '.git' folder in your '.stignore' file? [y/n] ")
+	ignoreGit, err := prompt.AskYesNo("Do you want to ignore the '.git' folder in your '.stignore' file? [y/n] ")
 	if err != nil {
 		return fmt.Errorf("failed to ask for adding '.git' to '%s': %s", stignorePath, err.Error())
 	}
